requirements.txt
django < 5.0.0
djangorestframework
psycopg2-binary

